requirements.txt
alembic==1.6.5
anyio==3.3.0
apache-airflow==2.1.2
apache-airflow-providers-ftp==2.0.0
apache-airflow-providers-imap==2.0.0
apache-airflow-providers-sqlite==2.0.0
apispec==3.3.2
appdirs==1.4.4
appnope==0.1.0
argcomplete==1.12.3
argon2-cffi==20.1.0
asn1crypto==1.4.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
autopep8==1.5.7
azure-common==1.1.27
azure-core==1.17.0
azure-storage-blob==12.8.1
Babel==2.9.1
backcall==0.2.0
bleach==3.2.1
blinker==1.4
boto3==1.18.21
botocore==1.21.21
cachetools==4.2.1
cattrs==1.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
croniter==1.0.15
cryptography==3.4.7
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.10
dill==0.3.4
distlib==0.3.1
dnspython==2.1.0
docutils==0.16
email-validator==1.1.3
entrypoints==0.3
Faker==13.3.2
filelock==3.0.12
Flask==1.1.4
Flask-AppBuilder==3.3.2
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gcloud==0.18.3
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.25.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud==0.34.0
google-cloud-bigquery==2.10.0
google-cloud-core==1.6.0
google-cloud-secret-manager==2.2.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
graphviz==0.17
grpc-google-iam-v1==0.12.3
grpcio==1.35.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.6
httplib2==0.19.0
httpx==0.19.0
idna==2.10
importlib-resources==1.5.0
inflection==0.5.1
install==1.3.4
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.16
isodate==0.6.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
libcst==0.3.17
lkml==0.2.2
lockfile==0.12.2
looker-sdk==0.1.3b20
lookml==3.0.3
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
msrest==0.6.21
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
ndg-httpsclient==0.5.1
nest-asyncio==1.4.3
nose==1.3.7
notebook==6.1.5
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
oscrypto==1.2.1
packaging==20.4
pandas==1.2.1
pandocfilters==1.4.3
parso==0.7.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prison==0.1.3
prometheus-client==0.9.0
prompt-toolkit==3.0.8
proto-plus==1.13.0
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodomex==3.10.1
PyGithub==1.53
Pygments==2.7.2
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-daemon==2.3.0
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.4
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole==4.7.7
QtPy==1.9.0
Random-Word==1.0.7
requests==2.25.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==10.7.0
rsa==4.7
s3transfer==0.5.0
scipy==1.8.0
seaborn==0.11.2
selenium==3.141.0
Send2Trash==1.5.0
setproctitle==1.2.2
six==1.15.0
sniffio==1.2.0
snowflake-connector-python==2.5.1
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
timedelta==2020.12.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
typing-inspect==0.6.0
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.4.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.3
zope.interface==5.4.0
